+++ src/main/kotlin/com/noumenadigital/npl/cli/config/DeployConfig.kt
@@ -22,15 +22,16 @@ sealed interface DeploymentTargetConfig
 data class EngineTargetConfig(
     val engineManagementUrl: String = "http://localhost:12400/realms/noumena",
     val authUrl: String = "http://localhost:11000",
-    val username: String = "",
-    val password: String = "",
-    val clientId: String = "foo",
-    val clientSecret: String = "bar",
+    val username: String,
+    val password: String,
+    val clientId: String? = null,
+    val clientSecret: String? = null,
 ) : DeploymentTargetConfig
 
 data class DeployConfig(
     val schemaVersion: String = "v1",
     val targets: Map<String, DeploymentTargetConfig> = emptyMap(),
+    val defaultTarget: String? = null,
 ) {
     companion object {
         private val mapper = ObjectMapper(YAMLFactory()).registerModule(KotlinModule.Builder().build())