+++ CHANGELOG.md
@@ -7,3 +7,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Added
+
+## [2025.1.5]
+
+Initial "new" CLI, with the following commands:
+
+- help
+- version
+- check
+- test
+- openapi
+- deploy
+
+### Removed
+
+All of the features of the old CLI (all related to cloud deployment).